requirements.txt
kivy>=2.3.0
kivymd>=1.1.1
cryptography>=41.0.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
opencv-python>=4.8.0
pyaudio>=0.2.11
pydub>=0.25.1
qrcode>=7.4.0
liboqs-python>=0.8.0